feat(hash_cache, #122): optimize doubly linked list memory usage
Previously, data for each entry for the LRU doubly linked list is
embedded in the value type, but this may cause each entry to use much
more memory than 2-byte when the size of the value type is proportional
to the alignment.

Now, the 2-byte data is stored in a dedicated space in the bucket, thus
improving memory locality and space efficiency.